CHanged time type

+++ src/core/Score.java
@@ -2,7 +2,7 @@ package core;
 
 public class Score {
 	private String name;
-	private int time;
+	private long time;
 	private Board.Dificulty difficulty;
 	
 	public Score(String name, int time, Board.Dificulty difficulty) {
@@ -25,7 +25,7 @@ public class Score {
 		return this.name;
 	}
 	
-	public int getTime() {
+	public long getTime() {
 		return this.time;
 	}
